+++ sc-types/src/lib.rs
@@ -255,9 +255,9 @@ pub enum GameCommand {
 #[derive(Deserialize, Serialize)]
 pub enum ClientPkt {
     Hello { seq: i32, sent_time: f64 },
-    Target { seq: i32, ack: i32, updates: VecDeque<(i64, Vec<GameCommand>)>, frame: i64, frame_ack: i64, frame_delay: u8 },
-    Ended { seq: i32, ack: i32, frame: i64 },
-    StateHash { seq: i32, ack: i32, hash: u32, frame: i64 }
+    Target { seq: i32, ack: i32, updates: VecDeque<(i32, Vec<GameCommand>)>, frame: i32, frame_ack: i32, frame_delay: u8 },
+    Ended { seq: i32, ack: i32, frame: i32 },
+    StateHash { seq: i32, ack: i32, hash: u32, frame: i32 }
 }
 
 #[derive(Deserialize, Serialize)]
@@ -272,5 +272,5 @@ pub struct ServerPkt {
 pub enum ServerEnum {
     Welcome { handshake_start_time: f64, player_id: usize },
     Start { rng_seed: [u8; 32] },
-    UpdateOtherTarget { updates: VecDeque<(i64, Vec<GameCommand>)>, frame: i64, frame_ack: i64, frame_delay: u8 }
+    UpdateOtherTarget { updates: VecDeque<(i32, Vec<GameCommand>)>, frame: i32, frame_ack: i32, frame_delay: u8 }
 }
